Set encoding of output file to be utf-8
By setting the encoding of the output file to be utf-8, it allows for emojis and other special characters found in the text to be written to the output file. Currently, a "UnicodeEncodeError: 'charmap' codec can't encode character..." error is thrown if attempting to run classify.py on text containing Unicode characters.